fix: actualizamos importaciones y organizamos modelo base
