+++ server/app/crud/user.py
@@ -1,6 +1,6 @@
 from sqlalchemy.ext.asyncio import AsyncSession
 from sqlalchemy.future import select
-from app.models import User
+from app.models.user import User
 from app.schemas.user import UserCreate
 from passlib.hash import bcrypt
